requirements.txt
streamlit==1.32.0
google-generativeai==0.7.2
python-dotenv
protobuf==4.25.3